+++ R/vcov3_fixest.R
@@ -116,61 +116,12 @@ calculate_beta_jack_dense <- function(obj, cluster, type, absorb_cluster_fixef, 
     cluster <- reformulate(cluster)
   }
 
-  cluster_tmp <-
-    try(
-      if("Formula" %in% loadedNamespaces()) { ## FIXME to suppress potential
-        #warnings due to | in Formula
-        suppressWarnings(expand.model.frame(
-          model = obj,
-          extras = cluster,
-          na.expand = FALSE,
-          envir = call_env
-        )
-        )
-      } else {
-        expand.model.frame(
-          obj,
-          cluster,
-          na.expand = FALSE,
-          envir = call_env
-        )
-      }
-    )
-
-  if (
-    inherits(
-      cluster_tmp,
-      "try-error"
-    ) &&
-    grepl(
-      "non-numeric argument to binary operator$",
-      attr(
-        cluster_tmp,
-        "condition"
-      )$message
-    )
-  ) {
-    cli::cli_abort(
-      "In your model, you have specified multiple fixed effects,
-      none of which are of type factor. While `fixest::feols()` handles
-      this case gracefully,  `summclust()` currently cannot handle this
-      case - please change the type of (at least one) fixed effect(s) to
-      factor. If this does not solve the error, please report the issue
-      at https://github.com/s3alfisc/summclust"
-    )
-  }
-
-  cluster_df <- model.frame(cluster, cluster_tmp, na.action = na.pass)
-  unique_clusters <- as.character(unique(cluster_df[, , drop = TRUE]))
-
-  G <- length(unique_clusters)
-  small_sample_correction <- (G-1)/G
-
-  N_g <- lapply(
-    seq_along(unique_clusters),
-    function(x) nrow(cluster_df[,,drop = TRUE] == x)
-  )
-
+  cluster_df <- get_cluster(
+    object = obj,
+    cluster = cluster,
+    N = N,
+    call_env = call_env
+  )$cluster_df
 
   k <- obj$nparams