+++ tests/testthat/test-methods.R
@@ -7,7 +7,7 @@ test_that("test methods", {
   # few large clusters (around 10000 obs)
   N <- 1000
   N_G1 <- 10
-  data <- summclust:::create_data(
+  data1 <<- summclust:::create_data(
     N = N,
     N_G1 = N_G1,
     icc1 = 0.8,
@@ -21,12 +21,12 @@ test_that("test methods", {
   feols_fit <- feols(
     proposition_vote ~ treatment + log_income | group_id1,
     cluster = ~group_id1,
-    data = data
+    data = data1
   )
 
   lm_fit <- lm(
     proposition_vote ~ treatment + log_income + group_id1,
-    data = data
+    data = data1
   )
 
   summ_feols <-