+++ NEWS.md
@@ -7,6 +7,12 @@
 
 # summclust 0.6
 
+# summclust 0.7
+
+* fixes a bug that turned NA values in cluster variables not contained in the model call into a distinct cluster. Instead, `summclust` now throws an error.
+
+# summclust 0.6
+
 * support for regression weights / weighted least squares
 * prettier errors and warnings via `cli` (0.5.2)